feat: Sent Module 모듈화된 파일 채택
