+++ Projects/Core/SSNetwork/Sources/SSNetworkTargetType.swift
@@ -68,4 +68,3 @@ enum DefaultHeaderValue: String, CaseIterable {
     }
   }
 }
-
