+++ Projects/Feature/Received/Sources/LedgerDetail/LedgerDetailEdit/LedgerDetailEdit.swift
@@ -6,9 +6,9 @@
 //  Copyright © 2024 com.oksusu. All rights reserved.
 //
 import ComposableArchitecture
+import Designsystem
 import FeatureAction
 import Foundation
-import Designsystem
 
 // MARK: - LedgerDetailEdit
