+++ Projects/Feature/Sent/Project.swift
@@ -15,7 +15,6 @@ let project = Project.makeModule(
       .feature(.sSSelectableItems),
       .feature(.sSCreateEnvelope),
       .feature(.sSEnvelope),
-      .feature(.sSEditSingleSelectButton),
     ],
     testDependencies: []
   )
